[Refactor] DraggableMap 컴포넌트 분리
